Results and link
Results
In this markdown I'm showing the progres of my model.

Iteration 0: Baseline
Goal: My Baseline is provided as an example, in which survivers are selected only by gender.

Train accuracy = 0.8922
Test accuracy = 0.7655
Conclusion: Whatever model which results are equal or worse than the baseline is almost worthless.

Iteration 1: Firt simple model
Goal: In this first model the goal is to create a structure to be improved, take into account some of the variables (Age is not included) and obtain a model to which incremental improvements are added.

Train accuracy = 0.8698
Test accuracy = 0.7727
Conclusion: It is a very small improvement over the Baseline. The model is still clearly overfitted, but I prefer to work including new variables, as Age, and then work in that direction.

Iteration 2: Data analysis
Goal: In this second model the goal is to do a previous treatment of the data that allows us to discard useless data, enhancing the useful ones and simplifying the model.

Train accuracy = 0.8720
Test accuracy = 0.7608
Conclusion: It is not an improvement over the Baseline. The model is still clearly overfitted. Let's work in the model.

Iteration 3: Different models
Goal: Lets try different models and see which one is better.

Train accuracy = 0.8307
Test accuracy = 0.7631
Conclusion: RandomTree was the best model but still is not an improvement over the Baseline. Let's keep thinking and improving.

Iteration 4: All models
Goal: Lets try different mixing the differente model and see if this is uselful.

Train accuracy = 0.8245
Test accuracy = 0.7656
Conclusion: There is almost an improvement over the Baseline. Let's try with the data again.

Iteration 5: New dat analysis
Goal: Get a better understanding of the data.

Train accuracy = 0.8193
Test accuracy = 0.7751
Conclusion: Best result so far with Logistic Model. Let's try with the data again.

Iteration 6: All models
Goal: Lets try different mixing the differente model and see if this is uselful.

Train accuracy = 0.8245
Test accuracy = 0.7775
Conclusion: Anothner small improvement.


Link:
https://www.kaggle.com/ignaciomorenomarn/titanic-prediction