Results and link
Results

In this markdown I'm showing the progres of my model.

Iteration 0: Baseline
Goal: My Baseline is the percentage of numbers that I would classification by chance. As all number are almost equally represented that proportion is 10%.

Validation accuracy = 0.10
Conclusion: Whatever model which results are equal or worse than the baseline is almost worthless.

Iteration 1: Firt simple model
Goal: In this first model the goal is to create a structure to be improved, take into account some of the variables and obtain a model to which incremental improvements are added.

Validation accuracy = 0.9872
Test accuracy = 0.9880
Conclusion: It is a very good result. I think it can improve if more epochs are allowed.

Iteration 2: More epochs
Goal: Lets try with more epochs.

Validation accuracy = 0.9883
Test accuracy = 0.9887
Conclusion: It is a small improvement, but the model is now much more slower. There is not much room for improvement in this area. Lets observate the mismatched data and look for a patron

Iteration 3: Searching in the mismatches
Goal: Identificate patrons in the mismatched data.

Conclusion: I couldn't identificate most of the numbers by myself so this competition is over for now.

Link
https://www.kaggle.com/ignaciomorenomarn/digit-recognition